change-structure: Separated components from containers
